Change html line separator from `</br>` to `<br>`
The correct HTML tag for a line break is simply `<br>`. The current implementation uses `</br>`, which would be the closing tag for a line break. Since the element does not accept any contents, there is no need for a closing tag. Also, the closing tag alone does not make any sense. There should be only the opening tag.

References:
- https://stackoverflow.com/questions/1946426/html-5-is-it-br-br-or-br
- https://www.w3schools.com/TAGS/tag_br.asp