changedetection/refactor: Improve init and close performance by removing unnecessary multiprocessing in root changedetection.py I assume that the purpose of existence of the root changedetection.py is for dev and container. But it doesn't mean to use multiprocess to start CD server.
I remove the multiprocess because of sevaral reasons below.

Benefit of this is that it is still the same $PWD of the previous root
changedetection.py. So things related with relative path is safe.
Also user can still develope and execute the root changedetection.py
in the root of the repo. Also It doesn't break Dockerfile CMD statement.
Also this change allows to use only one pid rather than two pids.

Conventional reason.
Also using SIGINT to the process with kill command directly seems abusing SIGINT signal.
For example, jupyter-notebook prevents capturing signal when isatty is
False. So jupyter-notebook won't accept SIGNINT if not connected with
a terminal.
https://github.com/jupyter/notebook/blob/8e7c99ad8d3fe59f98bb03e828307387a3694704/notebook/notebookapp.py#L1810
https://stackoverflow.com/a/55444671/20307768

Technical reason.
This is the slightly detailed SIGINT mechanism.
https://stackoverflow.com/a/31907519/20307768
https://unix.stackexchange.com/a/403605

So when you sent SIGINT to changedetectionio/changedetection.py,
root changedetection.py and changedetectionio/changedetection.py are received at the same time.
That is the reason I remove the signal catch message in root one.

But when you sent SIGINT to root changedetection.py(that invokes multiprocess),
It doesn't send it to the spawn process.
No matter what the reason behind it(Of course, most wrong idea I have
about unix is because I didn't know deep details and history.),
I tried to use pid(Process.pid) of the spwan one to send SIGINT using signal.pidfd_send_signal
however the method is available only Linux.(https://docs.python.org/3/library/signal.html#signal.pidfd_send_signal)
So in this case I needed to use IPC to get PID of the spawn process. But it is too much and only prolongs exiting time.