Modernize setuptools to use pyproject.toml
This is inline with the recommendations in Python Packaging User Guide at
https://packaging.python.org/en/latest/guides/modernize-setup-py-project/
Migration to the new format should not change any build steps since the
newer build invokations are already being used (e.g., "python -m build").

`requirements.txt` is left as-is and read as "dynamic" data (i.e., outside
of the `pyproject.toml` file) by setuptools. A further enhancement may
be to separate out the production requirements from the development
environment requirements if any exist.

`zip_safe` directive is removed since it is deprecated (See
https://setuptools.pypa.io/en/latest/deprecated/zip_safe.html).

`scripts` is also deprecated in favor of `entry_points` which is now
under `[projects.scripts]` in the same format. This means the
`changedetection.py` file is not included, but it appears to be intended
for local development as the `changedetection.io` script created during
installation from `entry_points` has the same function.

`include_package_data` is true by default in `pyproject.toml` format, so
it is excluded from the new file.

The changes were verified to build the same `.whl` and `.tar.gz` by
examining the artifacts. The exceptions are:
  * In the `.whl` file there is no longer `changedetection.py` in the
    `scripts` directory.
  * `zip_safe` is no longer included
  * `pyproject.toml` is now included
  * In `PKG-INFO`:
    * `Home-page` field is now in `Project-URL`
    * Additional Classifier for `Apache License 2.0`
    * `Keywords` are now keywords instead of a long string