Don't restrict .box to 80% width
Having `max-width: 80%` is very frustrating.

I can see my browser window has enough space for a wider table, and yet the table cells are squished while leaving empty whitespace at the sides.

This is specially noticeable when using a non-wide monitor and resizing the browser to half of the monitor width. (In my case, 2240x1400 / 2 = 1120x1400.)